telegram/requirements.txt
anyio==4.2.0
cachetools==5.3.2
certifi==2024.2.2
charset-normalizer==3.3.2
google-api-core==2.17.1
google-auth==2.27.0
google-cloud-speech==2.24.1
googleapis-common-protos==1.62.0
grpcio==1.60.1
grpcio-status==1.60.1
h11==0.14.0
httpcore==1.0.3
httpx==0.26.0
idna==3.6
proto-plus==1.23.0
protobuf==4.25.2
pyasn1==0.5.1
pyasn1-modules==0.3.0
python-telegram-bot==20.8
requests==2.31.0
rsa==4.9
sniffio==1.3.0
urllib3==2.2.0
